* cpdf_id_t changed to uintptr_t (portable type which is guaranteed to hold   pointer value without data loss) * CPdf::id is assigned with reinterpret_cast rather than simple cast because   reinterpret_cast is C++ clean way how to cast types with potentially different   data width

+++ src/kernel/cpdf.h
@@ -14,6 +14,8 @@
 
 #include "kernel/static.h"
 
+#include <inttypes.h>
+
 // xpdf
 #include "kernel/xpdf.h"
 
@@ -227,7 +229,7 @@ public:
 
 	/** Type for pdf identificator.
 	 */
-	typedef unsigned long long cpdf_id_t;
+	typedef uintptr_t cpdf_id_t;
 
 	/** Constant for pdf id of no pdf.
 	 * This is used for properties which comes from no pdf. Each CPdf instance
